improve build logic
Previously, Output#update() called Output#build() at a root node
before it returned the initiating input's updated hash.  As a result,
an input's updated hash was not reflected in an output computation
that depended on it.  The new logic finds the roots and calls
Output#build() on them after awaiting the input's hash update.